+++ app/src/main/java/com/example/moviesplanet/data/storage/remote/MoviesServiceApi.kt
@@ -1,6 +1,5 @@
 package com.example.moviesplanet.data.storage.remote
 
-import io.reactivex.Observable
 import io.reactivex.Single
 import retrofit2.http.GET
 import retrofit2.http.Path
@@ -9,7 +8,7 @@ import retrofit2.http.Query
 interface MoviesServiceApi {
 
     @GET("3/movie/{sort}")
-    fun getMovies(@Path("sort") sorted: String, @Query("page") page: Int): Single<MovieListResponse>
+    fun getMovies(@Path("sort") sorted: String, @Query("page") page: Long): Single<MovieListResponse>
 
     @GET("3/movie/{id}/videos")
     fun getMovieVideos(@Path("id") id: String): Single<VideoListResponse>
